+++ fed_lth/.gitignore
@@ -2,3 +2,4 @@
 /__pycache__
 /models
 /temp
+/conf.py
