+++ fed_lth/local_model.py
@@ -10,6 +10,7 @@ import time
 import torchvision
 from torchvision.models import resnet18
 import torch_pruning as tp
+from pruning_utils import MySlimmingImportance,MySlimmingPruner
 import torch.nn as nn
 import torch.optim as optim
 
@@ -18,7 +19,7 @@ class Local_model(object):
     self.id=id
     self.model=None  #默认为none 等待服务器下发模型
     # 本地数据集
-    self.train_data,self.test_data=get_dataset(id)
+    self.train_data,self.test_data, self.train_len,_=get_dataset(id)
 
   # 随机训练测试时间
   def time_test(self):
@@ -47,31 +48,12 @@ class Local_model(object):
 
     # 计算并打印时间
     elapsed_time = end_time - start_time
-    final_time = len(self.test_data)*elapsed_time/16
-
+    final_time = self.train_len*elapsed_time/16
+    print(f'final_time:{final_time}')
     return final_time
 
   def s_prune(self,ratio):
     # 剪枝函数 输入是剪枝率
-    class MySlimmingImportance(tp.importance.Importance):
-      def __call__(self, group, **kwargs):
-        group_imp = []
-        for dep, idxs in group:
-          layer = dep.target.module
-          prune_fn = dep.handler
-          if isinstance(layer, (nn.BatchNorm1d, nn.BatchNorm2d, nn.BatchNorm3d)) and layer.affine:
-            importance_scores = torch.abs(layer.weight.data)  # 使用 L1 范数计算重要性
-            group_imp.append(importance_scores)
-        if len(group_imp) == 0: return None
-        group_imp = torch.stack(group_imp, dim=0).mean(dim=0)
-        return group_imp
-
-    class MySlimmingPruner(tp.pruner.MetaPruner):
-      def regularize(self, model, reg):
-        for m in model.modules():
-          if isinstance(m, (nn.BatchNorm1d, nn.BatchNorm2d, nn.BatchNorm3d)) and m.affine:
-            m.weight.grad.data.add_(reg * torch.sign(m.weight.data))
-
     imp = MySlimmingImportance()
 
     # 忽略最后的分类层
@@ -92,23 +74,23 @@ class Local_model(object):
       ignored_layers=ignored_layers,
     )
 
-    # 定义损失函数和优化器（可更改）
-    criterion = nn.CrossEntropyLoss()
-    optimizer = optim.SGD(self.model.parameters(), lr=0.001, momentum=0.9)
+    # # 定义损失函数和优化器（可更改）
+    # criterion = nn.CrossEntropyLoss()
+    # optimizer = optim.SGD(self.model.parameters(), lr=0.001, momentum=0.9)
 
-    # 模拟训练数据
-    inputs = torch.randn(16, 3, 224, 224)  # 16张 224x224 的图片
-    labels = torch.randint(0, 1000, (16,))  # 16个标签，范围在0-999之间
+    # # 模拟训练数据
+    # inputs = torch.randn(16, 3, 224, 224)  # 16张 224x224 的图片
+    # labels = torch.randint(0, 1000, (16,))  # 16个标签，范围在0-999之间
 
-    # 进行稀疏训练
-    for epoch in range(5):  # 训练5个epoch
-      self.model.train()
-      optimizer.zero_grad()
-      outputs = self.model(inputs)
-      loss = criterion(outputs, labels)
-      loss.backward()
-      pruner.regularize(self.model, reg=1e-5)  # 稀疏化
-      optimizer.step()
+    # # 进行稀疏训练
+    # for epoch in range(5):  # 训练5个epoch
+    #   self.model.train()
+    #   optimizer.zero_grad()
+    #   outputs = self.model(inputs)
+    #   loss = criterion(outputs, labels)
+    #   loss.backward()
+    #   pruner.regularize(self.model, reg=1e-5)  # 稀疏化
+    #   optimizer.step()
 
     pruner.step()
 
@@ -117,13 +99,13 @@ class Local_model(object):
     model = self.model
     start = 0.0
     end = 1.0
-    step = 0.1
+    step = 0.03
     current = start
     while current <= end:
       self.model = model
       self.s_prune(current)
       if self.time_test() <= time_T:
-        self.model = model
+        self.model = torchvision.models.resnet18()
         prune_ratio = current
         break
       current += step
@@ -133,11 +115,11 @@ class Local_model(object):
 
 #测试代码写在这里面
 if __name__=='__main__':
-  local=Local_model(1)
+  local=Local_model(0)
   local.model=torchvision.models.resnet18()
   train_time=local.time_test()
   print(train_time)
-  time_T=1000
+  time_T=470
   prune_ratio=local.prune_ratio(time_T)
   print(prune_ratio)
 
@@ -149,4 +131,5 @@ if __name__=='__main__':
   # 剪枝后计算模型参数和计算量
   macs, nparams = tp.utils.count_ops_and_params(local.model, example_inputs)
   print("参数数量: {:.2f} M => {:.2f} M".format(base_nparams / 1e6, nparams / 1e6))
+  print("稀疏率: {:.2f}%".format((base_nparams - nparams )/base_nparams))
   print("计算量: {:.2f} G => {:.2f} G".format(base_macs / 1e9, macs / 1e9))
